Add renderCurrentWeatherCard function to render current weather card. Improve resuability, organization, and maintainability.
